+++ okapi/src/lib.rs
@@ -1,3 +1,6 @@
+#![forbid(unsafe_code)]
+#![deny(clippy::all)]
+
 pub type Map<K, V> = schemars::Map<K, V>;
 
 pub mod openapi3;